+++ src/main.cpp
@@ -4,6 +4,7 @@ using namespace geode::prelude;
 
 #include <Geode/modify/MenuLayer.hpp>
 #include <Geode/modify/CCKeyboardDispatcher.hpp>
+#include <Geode/modify/VideoOptionsLayer.hpp>
 
 #include "BongoCat.hpp"
 
@@ -13,8 +14,14 @@ class $modify(MenuLayer) {
 	bool init() {
 		if (!MenuLayer::init()) return false;
 
-		// $on_mod(Loaded) crashes
 		if(!added) {
+			auto runningScene = CCDirector::sharedDirector()->getRunningScene();
+			auto catNode = runningScene->getChildByType<BongoCat>(0);
+			if (catNode) {
+				catNode->removeFromParentAndCleanup(false);
+				SceneManager::get()->forget(catNode);
+			}
+
 			BongoCat::addCat();
 			added = true;
 		}
@@ -46,3 +53,15 @@ class $modify(CCKeyboardDispatcher) {
 		return true;
 	}
 };
+
+// Fix graphics getting messed up on fullscreen switch or res change
+#ifdef GEODE_IS_DESKTOP
+
+class $modify(VideoOptionsLayer) {
+	void onApply(CCObject * sender) {
+		VideoOptionsLayer::onApply(sender);
+		added = false;
+	}
+};
+
+#endif