+++ changelog.md
@@ -1,6 +1,8 @@
 # 1.1.0
 - Bongo Cat now also bongos when you jump using the keyboard
 - Added more hats and extra decorations
+- Fixed bug where graphics would get messed up on fullscreen switch or resolution change
+- Fixed bug where Bongo Cat would be offscreen on screen resolution change
 
 # 1.0.0
 - Release